refactor(hooks): collapse the two query-key markers into one module
ALL_PAGES_QUERY_KEY_MARKER and INFINITE_QUERY_KEY_MARKER solved the same
problem — keeping a query's cache entry separate from the generated
single-page one — from two modules with different visibility, one public
through the multi barrel and one internal to utils.

They are now ESI_QUERY_KEY_MARKER.infinite and .allPages, appended by a
shared markEsiQueryKey, with the reason all three entries are distinct
documented once. That documentation also records the consequence worth
knowing: infinite and all-pages do not share an entry, so using both the
single-subject infinite hook and the multi-subject hook for one endpoint
fetches that collection twice.

Marker values are unchanged, so no cache key moves.
