+++ packages/hooks/src/hooks/calendar/useCharacterCalendar.ts
@@ -6,7 +6,7 @@ import {
 } from "@jitaspace/esi-client";
 
 import { useAccessToken } from "../auth";
-import { esiInfiniteQueryKey } from "../utils/esiInfiniteQueryKey";
+import { esiInfiniteQueryKey } from "../utils/esiQueryKeys";
 
 export const useCharacterCalendar = (characterId?: number) => {
   const { accessToken, authHeaders } = useAccessToken({
