+++ packages/hooks/src/hooks/contacts/useCharacterContacts.ts
@@ -14,7 +14,7 @@ import {
 } from "@jitaspace/esi-client";
 
 import { useAccessToken } from "../auth";
-import { esiInfiniteQueryKey } from "../utils/esiInfiniteQueryKey";
+import { esiInfiniteQueryKey } from "../utils/esiQueryKeys";
 
 export type CharacterContact =
   GetCharactersCharacterIdContactsQueryResponse[number];
