+++ packages/hooks/src/hooks/mail/useCharacterMails.ts
@@ -9,7 +9,7 @@ import {
 } from "@jitaspace/esi-client";
 
 import { useAccessToken } from "../auth";
-import { esiInfiniteQueryKey } from "../utils/esiInfiniteQueryKey";
+import { esiInfiniteQueryKey } from "../utils/esiQueryKeys";
 
 export function useCharacterMails(characterId?: number, labels: number[] = []) {
   const { accessToken, authHeaders } = useAccessToken({
